% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anole}
\alias{anole}
\title{Dorsal pattern image of a female brown anole lizard}
\format{A data.table with 1675 rows and 2 variables:
\describe{
  \item{x}{x-coordinate of a pixel}
  \item{y}{y-coordinate of a pixel}
}}
\usage{
anole
}
\description{
A sample dorsal pattern pixel image of a female brown anole lizard traced from the ImageJ software.
}
\references{
Moon and Kamath (2016), Examining the Ecological, Morphological, and Behavioral Correlates of Dorsal Pattern Variations in Female Brown Anole Lizards (\emph{Anolis sagrei}).
}
\keyword{datasets}
